+++ semcore/base-components/src/components/animation/Collapse.tsx
@@ -6,7 +6,10 @@ import style from './style/keyframes.shadow.css';
 import { useForkRef } from '@semcore/core/lib/utils/ref';
 import useEnhancedEffect from '@semcore/core/lib/utils/use/useEnhancedEffect';
 
-function Collapse({ onAnimationStart, onAnimationEnd, overflowHidden = true, ...props }, ref) {
+function Collapse(
+  { onAnimationStart, onAnimationEnd, overflowHidden = true, defaultHeight = 'auto', ...props },
+  ref,
+) {
   const SCollapse = Animation;
   const overflowRef = React.useRef('initial');
   const innerRef = React.useRef(null);
@@ -22,7 +25,7 @@ function Collapse({ onAnimationStart, onAnimationEnd, overflowHidden = true, ...
 
   useEnhancedEffect(() => {
     if (!innerRef.current) return;
-    if (props.visible) innerRef.current.style.height = 'auto';
+    if (props.visible) innerRef.current.style.height = defaultHeight;
     if (!props.visible) innerRef.current.style.height = `${0}px`;
   }, []);
 
@@ -50,7 +53,7 @@ function Collapse({ onAnimationStart, onAnimationEnd, overflowHidden = true, ...
 
     setTimeout(() => {
       if (!element) return;
-      if (visibleRef.current) element.style.height = 'auto';
+      if (visibleRef.current) element.style.height = defaultHeight;
       if (!visibleRef.current) element.style.height = `${0}px`;
       if (overflowHidden) {
         element.style.overflow = overflowRef.current;