+++ semcore/data-table/src/components/Body/Row.types.ts
@@ -2,6 +2,7 @@ import { DTValue, DTUse, DataTableData } from '../DataTable/DataTable.types';
 import { DTColumn } from '../Head/Column.types';
 import { MergedColumnsCell, MergedRowsCell } from './MergedCells';
 import { ACCORDION, ROW_GROUP, ROW_INDEX, UNIQ_ROW_KEY } from '../DataTable/DataTable';
+import { DataTableCellProps } from './Cell.types';
 
 export type UniqRowKey = string;
 
@@ -17,6 +18,10 @@ export type DTRows = Array<DTRow | DTRow[]>;
 export type DataTableRowProps = {
   row: DTRow;
   mergedRow?: boolean;
+
+  isAccordionRow?: DataTableCellProps['isAccordionRow'];
+  animationExpand?: DataTableCellProps['animationExpand'];
+  accordionRowIndex?: DataTableCellProps['accordionRowIndex'];
 };
 
 export type RowPropsInner = JSX.IntrinsicElements['div'] & {
@@ -56,6 +61,7 @@ export type RowPropsInner = JSX.IntrinsicElements['div'] & {
 
   inert?: '';
 
+  accordionDuration?: number | [number, number];
   onBackFromAccordion: (colIndex: number) => void;
 
   scrollAreaRef: React.RefObject<HTMLDivElement>;