+++ semcore/base-components/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.0.1] - 2025-05-26
+
+### Fixed
+
+- `Delay` doesn't work for the `Collapse` component.
+
 ## [16.0.0] - 2025-05-19
 
 ### Added